refactor the package to use proper ts syntax - tslint should now also work as expected
