+++ src/asg-control.ts
@@ -1,85 +1,83 @@
-namespace angularSuperGallery {
+import {
+	IServiceController, IOptionsImage
+} from './asg-interfaces';
 
-	export class ControlController {
+export class ControlController {
 
-		public id: string;
-		private type = 'control';
-		private asg: IServiceController;
-		private template;
+	public id: string;
+	private type = 'control';
+	private asg: IServiceController;
+	private template;
 
-		constructor(
-			private service: IServiceController,
-			private $scope: ng.IScope) {
+	constructor(
+		private service: IServiceController,
+		private $scope: ng.IScope) {
 
-			this.template = 'views/asg-control.html';
-
-		}
-
-		public $onInit() {
-
-			// get service instance
-			this.asg = this.service.getInstance(this);
-
-			this.$scope.forward = () => {
-				this.asg.toForward(true);
-			};
-
-			this.$scope.backward = () => {
-				this.asg.toBackward(true);
-			};
-
-		}
-
-
-		// get image config
-		public get config(): IOptionsImage {
-
-			return this.asg.options[this.type];
-
-		}
-
-		// set image config
-		public set config(value: IOptionsImage) {
-
-			this.asg.options[this.type] = value;
-
-		}
-
-		// set selected image
-		public set selected(v: number) {
-
-			if (!this.asg) {
-				return;
-			}
-
-			this.asg.selected = v;
-
-		}
-
-		// get selected image
-		public get selected() {
-
-			if (!this.asg) {
-				return;
-			}
-
-			return this.asg.selected;
-
-		}
-
+		this.template = 'views/asg-control.html';
 
 	}
 
-	let app: ng.IModule = angular.module('angularSuperGallery');
+	public $onInit() {
 
-	app.component('asgControl', {
-		controller: ['asgService', '$scope', angularSuperGallery.ControlController],
-		template: '<div class="asg-control {{ $ctrl.asg.classes }}"><div ng-include="$ctrl.template"></div></div>',
-		bindings: {
-			id: '@?',
-			selected: '=?',
-			template: '@?'
+		// get service instance
+		this.asg = this.service.getInstance(this);
+
+		(<any>this.$scope).forward = () => {
+			this.asg.toForward(true);
+		};
+
+		(<any>this.$scope).backward = () => {
+			this.asg.toBackward(true);
+		};
+
+	}
+
+
+	// get image config
+	public get config(): IOptionsImage {
+
+		return this.asg.options[this.type];
+
+	}
+
+	// set image config
+	public set config(value: IOptionsImage) {
+
+		this.asg.options[this.type] = value;
+
+	}
+
+	// set selected image
+	public set selected(v: number) {
+
+		if (!this.asg) {
+			return;
 		}
-	});
+
+		this.asg.selected = v;
+
+	}
+
+	// get selected image
+	public get selected() {
+
+		if (!this.asg) {
+			return;
+		}
+
+		return this.asg.selected;
+
+	}
+
 
 }
+
+export const asgControlComponent = {
+	controller: ['asgService', '$scope', ControlController],
+	template: '<div class="asg-control {{ $ctrl.asg.classes }}"><div ng-include="$ctrl.template"></div></div>',
+	bindings: {
+		id: '@?',
+		selected: '=?',
+		template: '@?'
+	}
+};
