+++ src/asg.ts
@@ -1,31 +1,25 @@
-///<reference path="./../typings/index.d.ts" />
+import { module, IModule } from 'angular';
+import { asgBytesFilter } from './asg-bytes';
+import { asgControlComponent } from './asg-control';
+import { asgService } from './asg-service';
+import { asgImageComponent } from './asg-image';
+import { asgInfoComponent } from './asg-info';
+import { asgModalComponent } from './asg-modal';
+import { asgPanelComponent } from './asg-panel';
+import { asgThumbnailComponent } from './asg-thumbnail';
 
-namespace angularSuperGallery {
+export const moduleName = 'angularSuperGallery';
 
-	let app : ng.IModule = angular.module('angularSuperGallery', ['ngAnimate', 'ngTouch']);
-
-	app.filter('asgBytes', () => {
-		return function (bytes : any, precision : number) : string {
-
-			if (isNaN(parseFloat(bytes)) || !isFinite(bytes)) {
-				return '';
-			}
-
-			if (bytes === 0) {
-				return '0';
-			}
-
-			if (typeof precision === 'undefined') {
-				precision = 1;
-			}
-
-			let units = ['bytes', 'kB', 'MB', 'GB', 'TB', 'PB'],
-				number = Math.floor(Math.log(bytes) / Math.log(1024));
-
-			return (bytes / Math.pow(1024, Math.floor(number))).toFixed(precision) + ' ' + units[number];
-
-		};
-	});
-
-}
+export const app: IModule = module(moduleName, [
+	'ngAnimate',
+	'ngTouch'
+]);
+app.filter('asgBytes', asgBytesFilter);
+app.component('asgControl', asgControlComponent);
+app.service('asgService', asgService);
+app.component('asgImage', asgImageComponent);
+app.component('asgInfo', asgInfoComponent);
+app.component('asgModal', asgModalComponent);
+app.component('asgPanel', asgPanelComponent);
+app.component('asgThumbnail', asgThumbnailComponent);
